requirements.txt
fastapi==0.98.0
hypercorn==0.14.3
openai==0.27.8
python-dotenv==1.0.0
httpx==0.24.1